drop support for Go 1.13
