+++ .circleci/config.yml
@@ -5,11 +5,6 @@ executors:
       - image: "circleci/golang:1.14"
     environment:
       runrace: true
-  test-go113:
-    docker:
-      - image: "circleci/golang:1.13"
-    environment:
-      runrace: true
   interop:
     docker:
       - image: circleci/buildpack-deps:stretch
@@ -50,9 +45,6 @@ jobs:
           command: ginkgo -v -randomizeAllSpecs -trace integrationtests/self -- -qlog
   go114:
     <<: *test
-  go113:
-    <<: *test
-    executor: test-go113
   interop-build:
     executor: interop
     steps:
@@ -88,7 +80,6 @@ workflows:
   workflow:
     jobs:
       - go114
-      - go113
       - interop-build:
           filters:
             branches:
