+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/lucas-clemente/quic-go
 
-go 1.13
+go 1.14
 
 require (
 	github.com/cheekybits/genny v1.0.0
@@ -8,7 +8,7 @@ require (
 	github.com/golang/mock v1.4.0
 	github.com/golang/protobuf v1.4.0
 	github.com/marten-seemann/qpack v0.1.0
-	github.com/marten-seemann/qtls v0.9.1
+	github.com/marten-seemann/qtls v0.9.2
 	github.com/onsi/ginkgo v1.11.0
 	github.com/onsi/gomega v1.8.1
 	golang.org/x/crypto v0.0.0-20200423211502-4bdfaf469ed5
